ver 2.0 step 1

+++ api/repositories/user_repository.py
@@ -1,6 +1,7 @@
 from api.models.user import User
 from api import db
 
+
 class UserRepository:
 
     @staticmethod
@@ -8,26 +9,41 @@ class UserRepository:
         return User.query.all()
 
     @staticmethod
-    def get_user_by_id(user_id):
-        return User.query.get(user_id)
+    def get_user_by_profile_ref(profile_ref):
+        return User.query.filter_by(profile_ref=profile_ref).first()
 
     @staticmethod
-    def create_user(username):
-        user = User(username=username)
+    def create_user(user):
+        if not isinstance(user, User):
+            raise TypeError("Expected a User instance.")
+
+        if not user.profile_ref:
+            raise ValueError("User must have a profile_ref.")
+
         db.session.add(user)
         db.session.commit()
         return user
 
     @staticmethod
-    def update_user(user_id, username=None):
-        user = User.query.get(user_id)
-        if username:
-            user.username = username
+    def update_user(user):
+
+        if not isinstance(user, User):
+            raise TypeError("Expected a User instance.")
+
+        existing_user = User.query.filter_by(profile_ref=user.profile_ref).first()
+
+        if not existing_user:
+            return None  # Return if the user with the given profile_ref is not found
+
+        existing_user.username = user.username
+        existing_user.last_updated = user.last_updated
+
         db.session.commit()
-        return user
+        return existing_user
 
     @staticmethod
-    def delete_user(user_id):
-        user = User.query.get(user_id)
-        db.session.delete(user)
-        db.session.commit()
+    def delete_user(profile_ref):
+        user = User.query.filter_by(profile_ref=profile_ref).first()
+        if user:
+            db.session.delete(user)
+            db.session.commit()
